apps/stream-url-updater/pyproject.toml
[project]
name = "streamurlupdater"
version = "0.1.0"
description = "Generate m3u8 links"
requires-python = ">=3.10"
dependencies = [
    "hydra-core>=1.3.2",
    "selenium>=4.27.1",
]
